+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `SetOutput` method for setting an output writer on the Configurator
 
+### Changed
+
+- Improved usage output: added environment variables
+
 
 ## [0.4.0] - 2018-01-02